Remove extra classes & refactor
